refactor(runtime): extract compactor into pkg/runtime/compactor
Move the LLM-summarization strategy and session-aware extraction
helpers out of the runtime into a new pkg/runtime/compactor package.
The runtime keeps only what's intrinsically runtime-private: deciding
when to compact (loop.go), dispatching hooks, and applying the
produced summary to the session.

The shape of the change:

- pkg/compaction (pure logic, no session/agent deps) gains
  SplitIndexForKeep and FirstIndexInBudget — moved verbatim from the
  runtime's splitIndexForKeep / firstMessageToKeep with their
  user/assistant boundary-snap semantics intact.

- pkg/runtime/compactor (new) owns the session-aware bits:
  ExtractMessages, ComputeFirstKeptEntry, MapToSessionIndex, and the
  default LLM strategy RunLLM. The strategy receives the runtime as
  a RunAgent callback rather than importing pkg/runtime, so the
  dependency direction stays one-way (runtime → compactor).
  MaxKeepTokens and MaxSummaryTokens move here as exported constants
  so user-supplied strategies can match the runtime's policy.

- pkg/runtime/session_compaction.go shrinks from 197 LoC of mixed
  orchestration + LLM call + extraction helpers to ~130 LoC of
  pure orchestration: lookup contextLimit, dispatch
  before_compaction, choose strategy (hook-supplied or default LLM),
  apply the summary, emit events, dispatch after_compaction. The
  produceSummary helper makes the hook-vs-default branching obvious
  at a glance.

Tests follow ownership: TestExtractMessages*, TestComputeFirstKeptEntry
and TestMapToSessionIndex move to pkg/runtime/compactor;
TestSplitIndexForKeep and a new TestFirstIndexInBudget land in
pkg/compaction. The runtime keeps only the integration tests
(TestDoCompact*, TestSessionGetMessages_*) and the gold-standard
TestCompaction continues to exercise the full LLM-driven path
through loop.go.

Behaviorally: zero changes. TestCompaction,
TestCompactionOverflowDoesNotLoop and
TestDoCompactNoHooksMatchesPriorBehavior all pass — the
sub-runtime-with-WithSessionCompaction(false) construction still
lives in runCompactionAgent, so the existing recursion guard is
unchanged.

What strategy authors gain: a public compactor.RunLLM /
compactor.LLMArgs / compactor.RunAgent surface they can call from
a custom before_compaction hook to delegate back to the default
strategy with tweaks (different prompt, different model) instead of
copying ~40 lines of LLM-call boilerplate.
