+++ pkg/runtime/session_compaction_test.go
@@ -2,220 +2,24 @@ package runtime
 
 import (
 	"os"
-	"strings"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/docker/docker-agent/pkg/agent"
 	"github.com/docker/docker-agent/pkg/chat"
-	"github.com/docker/docker-agent/pkg/compaction"
 	"github.com/docker/docker-agent/pkg/config/latest"
 	"github.com/docker/docker-agent/pkg/session"
 	"github.com/docker/docker-agent/pkg/team"
 )
 
-func TestExtractMessagesToCompact(t *testing.T) {
-	newMsg := func(role chat.MessageRole, content string) session.Item {
-		return session.NewMessageItem(&session.Message{
-			Message: chat.Message{Role: role, Content: content},
-		})
-	}
-
-	tests := []struct {
-		name                     string
-		messages                 []session.Item
-		contextLimit             int64
-		additionalPrompt         string
-		wantConversationMsgCount int
-	}{
-		{
-			name:                     "empty session returns system and user prompt only",
-			messages:                 nil,
-			contextLimit:             100_000,
-			wantConversationMsgCount: 0,
-		},
-		{
-			name: "system messages are filtered out",
-			messages: []session.Item{
-				newMsg(chat.MessageRoleSystem, "system instruction"),
-				newMsg(chat.MessageRoleUser, "hello"),
-				newMsg(chat.MessageRoleAssistant, "hi"),
-			},
-			contextLimit:             100_000,
-			wantConversationMsgCount: 2,
-		},
-		{
-			name: "messages fit within context limit",
-			messages: []session.Item{
-				newMsg(chat.MessageRoleUser, "msg1"),
-				newMsg(chat.MessageRoleAssistant, "msg2"),
-				newMsg(chat.MessageRoleUser, "msg3"),
-				newMsg(chat.MessageRoleAssistant, "msg4"),
-			},
-			contextLimit:             100_000,
-			wantConversationMsgCount: 4,
-		},
-		{
-			name: "truncation when context limit is very small",
-			messages: []session.Item{
-				newMsg(chat.MessageRoleUser, "first message with lots of content that takes tokens"),
-				newMsg(chat.MessageRoleAssistant, "first response with lots of content that takes tokens"),
-				newMsg(chat.MessageRoleUser, "second message"),
-				newMsg(chat.MessageRoleAssistant, "second response"),
-			},
-			// Set context limit so small that after subtracting maxSummaryTokens + prompt overhead,
-			// not all messages fit.
-			contextLimit:             maxSummaryTokens + 50,
-			wantConversationMsgCount: 0,
-		},
-		{
-			name: "additional prompt is appended",
-			messages: []session.Item{
-				newMsg(chat.MessageRoleUser, "hello"),
-			},
-			contextLimit:             100_000,
-			additionalPrompt:         "focus on code quality",
-			wantConversationMsgCount: 1,
-		},
-		{
-			name: "cost and cache control are cleared",
-			messages: []session.Item{
-				session.NewMessageItem(&session.Message{
-					Message: chat.Message{
-						Role:         chat.MessageRoleUser,
-						Content:      "hello",
-						Cost:         1.5,
-						CacheControl: true,
-					},
-				}),
-			},
-			contextLimit:             100_000,
-			wantConversationMsgCount: 1,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			sess := session.New(session.WithMessages(tt.messages))
-
-			a := agent.New("test", "test prompt")
-			result, _ := extractMessagesToCompact(sess, a, tt.contextLimit, tt.additionalPrompt, time.Now)
-
-			assert.GreaterOrEqual(t, len(result), tt.wantConversationMsgCount+2)
-			assert.Equal(t, chat.MessageRoleSystem, result[0].Role)
-			assert.Equal(t, compaction.SystemPrompt, result[0].Content)
-
-			last := result[len(result)-1]
-			assert.Equal(t, chat.MessageRoleUser, last.Role)
-			expectedPrompt := compaction.UserPrompt
-			if tt.additionalPrompt != "" {
-				expectedPrompt += "\n\n" + tt.additionalPrompt
-			}
-			assert.Equal(t, expectedPrompt, last.Content)
-
-			// Conversation messages are all except first (system) and last (user prompt)
-			assert.Equal(t, tt.wantConversationMsgCount, len(result)-2)
-
-			// Verify cost and cache control are cleared on conversation messages
-			for i := 1; i < len(result)-1; i++ {
-				assert.Zero(t, result[i].Cost)
-				assert.False(t, result[i].CacheControl)
-			}
-		})
-	}
-}
-
-func TestSplitIndexForKeep(t *testing.T) {
-	msg := func(role chat.MessageRole, content string) chat.Message {
-		return chat.Message{Role: role, Content: content}
-	}
-
-	tests := []struct {
-		name      string
-		messages  []chat.Message
-		maxTokens int64
-		wantSplit int // expected split index
-	}{
-		{
-			name:      "empty messages",
-			messages:  nil,
-			maxTokens: 1000,
-			wantSplit: 0,
-		},
-		{
-			name: "all messages fit in keep budget - compact everything",
-			messages: []chat.Message{
-				msg(chat.MessageRoleUser, "short"),
-				msg(chat.MessageRoleAssistant, "short"),
-			},
-			maxTokens: 100_000,
-			wantSplit: 2, // all fit → compact everything
-		},
-		{
-			name: "recent messages kept, older ones compacted",
-			messages: []chat.Message{
-				msg(chat.MessageRoleUser, strings.Repeat("a", 40000)),      // ~10005 tokens
-				msg(chat.MessageRoleAssistant, strings.Repeat("b", 40000)), // ~10005 tokens
-				msg(chat.MessageRoleUser, strings.Repeat("c", 40000)),      // ~10005 tokens
-				msg(chat.MessageRoleAssistant, strings.Repeat("d", 40000)), // ~10005 tokens
-				msg(chat.MessageRoleUser, strings.Repeat("e", 40000)),      // ~10005 tokens
-				msg(chat.MessageRoleAssistant, strings.Repeat("f", 40000)), // ~10005 tokens
-			},
-			maxTokens: 20_100, // enough for exactly 2 messages
-			wantSplit: 4,      // last 2 messages are kept
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := splitIndexForKeep(tt.messages, tt.maxTokens)
-			assert.Equal(t, tt.wantSplit, got)
-		})
-	}
-}
-
-func TestExtractMessagesToCompact_KeepsRecentMessages(t *testing.T) {
-	// Create a session with many messages, some large enough that the last
-	// ~20k tokens are kept aside.
-	var items []session.Item
-	for range 10 {
-		items = append(items, session.NewMessageItem(&session.Message{
-			Message: chat.Message{
-				Role:    chat.MessageRoleUser,
-				Content: strings.Repeat("x", 20000), // ~5k tokens each
-			},
-		}), session.NewMessageItem(&session.Message{
-			Message: chat.Message{
-				Role:    chat.MessageRoleAssistant,
-				Content: strings.Repeat("y", 20000), // ~5k tokens each
-			},
-		}))
-	}
-
-	sess := session.New(session.WithMessages(items))
-	a := agent.New("test", "test prompt")
-
-	result, firstKeptEntry := extractMessagesToCompact(sess, a, 200_000, "", time.Now)
-
-	// The kept messages should not appear in the compaction result
-	// (only system + compacted messages + user prompt).
-	// Total: 20 messages × ~5k tokens = ~100k tokens.
-	// Keep budget: 20k tokens → ~4 messages kept.
-	// So compacted messages should be 20 - 4 = 16.
-	compactedMsgCount := len(result) - 2 // minus system and user prompt
-	assert.Less(t, compactedMsgCount, 20, "some messages should have been kept aside")
-	assert.Positive(t, compactedMsgCount, "some messages should be compacted")
-
-	// firstKeptEntry should point into sess.Messages
-	assert.Positive(t, firstKeptEntry, "firstKeptEntry should be > 0")
-	assert.Less(t, firstKeptEntry, len(sess.Messages), "firstKeptEntry should be within bounds")
-}
-
+// TestSessionGetMessages_WithFirstKeptEntry covers the session-side
+// reconstruction of a compacted conversation: when a Summary item has
+// FirstKeptEntry set, the session's GetMessages must surface the summary
+// followed by the kept tail. This is what makes the compactor's
+// FirstKeptEntry actually work in the next LLM turn.
 func TestSessionGetMessages_WithFirstKeptEntry(t *testing.T) {
-	// Build a session with some messages, then add a summary with FirstKeptEntry.
 	items := []session.Item{
 		session.NewMessageItem(&session.Message{
 			Message: chat.Message{Role: chat.MessageRoleUser, Content: "m1"},
@@ -246,7 +50,6 @@ func TestSessionGetMessages_WithFirstKeptEntry(t *testing.T) {
 
 	messages := sess.GetMessages(a)
 
-	// Extract just the non-system messages
 	var conversationMessages []chat.Message
 	for _, msg := range messages {
 		if msg.Role != chat.MessageRoleSystem {
@@ -254,15 +57,16 @@ func TestSessionGetMessages_WithFirstKeptEntry(t *testing.T) {
 		}
 	}
 
-	// Should have: summary (as user message), m4, m5
 	require.Len(t, conversationMessages, 3, "expected summary + 2 kept messages")
 	assert.Contains(t, conversationMessages[0].Content, "Session Summary:")
 	assert.Equal(t, "m4", conversationMessages[1].Content)
 	assert.Equal(t, "m5", conversationMessages[2].Content)
 }
 
+// TestSessionGetMessages_SummaryWithoutFirstKeptEntry pins backward
+// compatibility: a summary with no FirstKeptEntry must still work, with
+// the conversation continuing from messages that follow the summary item.
 func TestSessionGetMessages_SummaryWithoutFirstKeptEntry(t *testing.T) {
-	// Backward compatibility: summary without FirstKeptEntry should work as before.
 	items := []session.Item{
 		session.NewMessageItem(&session.Message{
 			Message: chat.Message{Role: chat.MessageRoleUser, Content: "m1"},
@@ -288,35 +92,9 @@ func TestSessionGetMessages_SummaryWithoutFirstKeptEntry(t *testing.T) {
 		}
 	}
 
-	// Should have: summary + m3 (messages after the summary)
 	require.Len(t, conversationMessages, 2)
 	assert.Contains(t, conversationMessages[0].Content, "Session Summary:")
 	assert.Equal(t, "m3", conversationMessages[1].Content)
-}
-
-// TestComputeFirstKeptEntry covers the helper used by the hook-supplied
-// summary path to keep the same tail-keep policy (maxKeepTokens) as the
-// LLM path.
-func TestComputeFirstKeptEntry(t *testing.T) {
-	a := agent.New("test", "")
-
-	t.Run("empty session returns 0", func(t *testing.T) {
-		sess := session.New()
-		assert.Equal(t, 0, computeFirstKeptEntry(sess, a))
-	})
-
-	t.Run("system messages don't shift the index", func(t *testing.T) {
-		// A short conversation: all messages fit in the keep budget, so
-		// the helper returns len(sess.Messages) — i.e. compact everything.
-		sess := session.New(session.WithMessages([]session.Item{
-			session.NewMessageItem(&session.Message{Message: chat.Message{Role: chat.MessageRoleSystem, Content: "sys"}}),
-			session.NewMessageItem(&session.Message{Message: chat.Message{Role: chat.MessageRoleUser, Content: "hi"}}),
-			session.NewMessageItem(&session.Message{Message: chat.Message{Role: chat.MessageRoleAssistant, Content: "hello"}}),
-		}))
-		// Whole conversation is short → split at end → all compacted.
-		got := computeFirstKeptEntry(sess, a)
-		assert.Equal(t, len(sess.Messages), got)
-	})
 }
 
 // TestDoCompactBeforeHookDeniesSkipsCompaction verifies that a
@@ -343,7 +121,6 @@ func TestDoCompactBeforeHookDeniesSkipsCompaction(t *testing.T) {
 	)
 	require.NoError(t, err)
 
-	// Pre-populate a session with a couple of messages, no summary item.
 	sess := session.New(session.WithMessages([]session.Item{
 		session.NewMessageItem(&session.Message{Message: chat.Message{Role: chat.MessageRoleUser, Content: "hi"}}),
 		session.NewMessageItem(&session.Message{Message: chat.Message{Role: chat.MessageRoleAssistant, Content: "hello"}}),
@@ -373,8 +150,8 @@ func TestDoCompactBeforeHookDeniesSkipsCompaction(t *testing.T) {
 }
 
 // TestDoCompactBeforeHookSuppliesSummary verifies that a
-// before_compaction hook returning HookSpecificOutput.Summary causes the
-// runtime to apply that summary verbatim and to skip the LLM-based
+// before_compaction hook returning HookSpecificOutput.Summary causes
+// the runtime to apply that summary verbatim and to skip the LLM-based
 // summarization (no new model call).
 func TestDoCompactBeforeHookSuppliesSummary(t *testing.T) {
 	const customSummary = "custom hook-supplied summary"
@@ -386,8 +163,8 @@ func TestDoCompactBeforeHookSuppliesSummary(t *testing.T) {
 		},
 	}
 
-	// The provider must NOT be called — if it is, we'll consume from the
-	// (empty) mockStream and the test will catch it.
+	// The provider must NOT be called — if it is, we'll consume from
+	// the (empty) mockStream and the test will catch it.
 	prov := &mockProvider{id: "test/mock-model", stream: &mockStream{}}
 	root := agent.New("root", "test",
 		agent.WithModel(prov),
@@ -432,7 +209,6 @@ func TestDoCompactBeforeHookSuppliesSummary(t *testing.T) {
 	assert.Equal(t, 1, compactionStartCount, "expected exactly one compaction-started event")
 	assert.Equal(t, 1, compactionDoneCount, "expected exactly one compaction-completed event")
 
-	// The session must have a Summary item appended with the hook-supplied text.
 	last := sess.Messages[len(sess.Messages)-1]
 	assert.Equal(t, customSummary, last.Summary,
 		"the session must record the hook-supplied summary as its last item")
@@ -440,12 +216,10 @@ func TestDoCompactBeforeHookSuppliesSummary(t *testing.T) {
 		"hook-supplied summaries cost nothing — no LLM was called")
 }
 
-// TestDoCompactAfterHookFires verifies that after_compaction fires when
-// a summary was applied (LLM-path or hook-path), and that the hook
-// receives the produced summary text.
+// TestDoCompactAfterHookFires verifies that after_compaction fires
+// when a summary was applied (LLM-path or hook-path), and that the
+// hook receives the produced summary text.
 func TestDoCompactAfterHookFires(t *testing.T) {
-	// A dedicated file lets us assert that the hook actually ran and
-	// observe the summary it received.
 	dir := t.TempDir()
 	logFile := dir + "/after.log"
 
@@ -457,7 +231,6 @@ func TestDoCompactAfterHookFires(t *testing.T) {
 			{Type: "command", Command: "echo '" + beforeJSON + "'", Timeout: 5},
 		},
 		AfterCompaction: []latest.HookDefinition{
-			// jq extracts the summary field and writes it to the log file.
 			{Type: "command", Command: "cat | jq -r '.summary' > " + logFile, Timeout: 5},
 		},
 	}
@@ -496,8 +269,6 @@ func TestDoCompactAfterHookFires(t *testing.T) {
 // emit the same SessionCompaction started/completed pair that all
 // existing UIs depend on.
 func TestDoCompactNoHooksMatchesPriorBehavior(t *testing.T) {
-	// Use a queueProvider so the LLM-based compaction can run without
-	// blocking. A single short summary stream is enough.
 	summaryStream := newStreamBuilder().
 		AddContent("summary").
 		AddStopWithUsage(1, 1).
